third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpecularLightingElement-dom-suraceScale-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'surfaceScale' attribute of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS specularElement.getAttribute('surfaceScale') is "1"
PASS specularElement.getAttribute('surfaceScale') is "10"
PASS successfullyParsed is true

TEST COMPLETE
